+++ .storybook/welcome-page/components-demo/demo/demo.style.js
@@ -1,8 +1,8 @@
-import styled from 'styled-components';
+import styled from "styled-components";
 
 export const ComponentShowcaseWrapper = styled.div`
   align-items: center;
-  background-image: linear-gradient(180deg,#f3f8fe,#fdf5f5);
+  background: #fafbfb;
   display: flex;
   flex-direction: column;
   padding: 120px 0 0;