+++ src/components/box/box.component.tsx
@@ -8,6 +8,8 @@ import {
   flexbox,
   FlexboxProps,
   ColorProps,
+  position,
+  PositionProps,
 } from "styled-system";
 import BaseTheme from "../../style/themes/base";
 import styledColor from "../../style/utils/color";
@@ -21,7 +23,8 @@ export interface BoxProps
   extends SpaceProps,
     LayoutProps,
     FlexboxProps,
-    ColorProps {
+    ColorProps,
+    Omit<PositionProps, "zIndex"> {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   as?: keyof JSX.IntrinsicElements | React.ComponentType<any>;
   /** String to set Box content break strategy. Note "anywhere" is not supported in Safari */
@@ -30,12 +33,15 @@ export interface BoxProps
   scrollVariant?: ScrollVariant;
   /** set the box-sizing attribute of the Box component */
   boxSizing?: BoxSizing;
+  /** Allows a tabindex to be specified */
+  tabIndex?: number | string;
 }
 
 const Box = styled.div<BoxProps>`
   ${space}
   ${layout}
   ${flexbox}
+  ${position}
   ${({ color, bg, backgroundColor, ...rest }) =>
     styledColor({ color, bg, backgroundColor, ...rest })}