+++ src/components/sidebar/sidebar.component.js
@@ -1,16 +1,21 @@
 import React, { useRef } from "react";
 import PropTypes from "prop-types";
+import styledSystemPropTypes from "@styled-system/prop-types";
 
 import Modal from "../modal";
-import SidebarStyle from "./sidebar.style";
+import StyledSidebar from "./sidebar.style";
 import IconButton from "../icon-button";
 import Icon from "../icon";
 import FocusTrap from "../../__internal__/focus-trap";
 import SidebarHeader from "./__internal__/sidebar-header";
 import Box from "../box";
-import { SIDEBAR_SIZES, SIDEBAR_ALIGNMENTS } from "./sidebar.config";
 import createGuid from "../../__internal__/utils/helpers/guid";
 import useLocale from "../../hooks/__internal__/useLocale";
+import { filterStyledSystemPaddingProps } from "../../style/utils";
+
+const paddingPropTypes = filterStyledSystemPaddingProps(
+  styledSystemPropTypes.space
+);
 
 export const SidebarContext = React.createContext({});
 
@@ -21,11 +26,11 @@ const Sidebar = React.forwardRef(
       "aria-label": ariaLabel,
       "aria-labelledby": ariaLabelledBy,
       open,
-      disableEscKey,
-      enableBackgroundUI,
+      disableEscKey = false,
+      enableBackgroundUI = false,
       header,
-      position,
-      size,
+      position = "right",
+      size = "medium",
       children,
       onCancel,
       role = "dialog",
@@ -37,7 +42,7 @@ const Sidebar = React.forwardRef(
     const locale = useLocale();
     const { current: titleId } = useRef(createGuid());
 
-    let sidebarRef = useRef();
+    let sidebarRef = useRef(null);
     if (ref) sidebarRef = ref;
     const closeIcon = () => {
       if (!onCancel) return null;
@@ -59,7 +64,7 @@ const Sidebar = React.forwardRef(
     };
 
     const sidebar = (
-      <SidebarStyle
+      <StyledSidebar
         aria-modal={!enableBackgroundUI}
         aria-describedby={ariaDescribedBy}
         aria-label={ariaLabel}
@@ -77,8 +82,10 @@ const Sidebar = React.forwardRef(
         {closeIcon()}
         <Box
           data-element="sidebar-content"
-          p={4}
-          pt="27px"
+          pt="var(--spacing300)"
+          pb="var(--spacing400)"
+          px="var(--spacing400)"
+          {...filterStyledSystemPaddingProps(rest)}
           scrollVariant="light"
           overflow="auto"
           flex="1"
@@ -87,7 +94,7 @@ const Sidebar = React.forwardRef(
             {children}
           </SidebarContext.Provider>
         </Box>
-      </SidebarStyle>
+      </StyledSidebar>
     );
 
     return (
@@ -116,6 +123,8 @@ const Sidebar = React.forwardRef(
 );
 
 Sidebar.propTypes = {
+  /** Styled system padding props to apply to Sidebar content */
+  ...paddingPropTypes,
   /** Prop to specify the aria-describedby property of the component */
   "aria-describedby": PropTypes.string,
   /** Prop to specify the aria-label of the component */
@@ -133,9 +142,17 @@ Sidebar.propTypes = {
   /** Set this prop to false to hide the translucent background when the dialog is open. */
   enableBackgroundUI: PropTypes.bool,
   /** Sets the position of sidebar, either left or right. */
-  position: PropTypes.oneOf(SIDEBAR_ALIGNMENTS),
+  position: PropTypes.oneOf(["left", "right"]),
   /** Sets the size of the sidebar when open. */
-  size: PropTypes.oneOf(SIDEBAR_SIZES),
+  size: PropTypes.oneOf([
+    "extra-small",
+    "small",
+    "medium-small",
+    "medium",
+    "medium-large",
+    "large",
+    "extra-large",
+  ]),
   /** Node that will be used as sidebar header. */
   header: PropTypes.node,
   /** The ARIA role to be applied to the container */
@@ -144,12 +161,6 @@ Sidebar.propTypes = {
   focusableContainers: PropTypes.arrayOf(
     PropTypes.shape({ current: PropTypes.any })
   ),
-};
-
-Sidebar.defaultProps = {
-  position: "right",
-  size: "medium",
-  enableBackgroundUI: false,
 };
 
 export default Sidebar;